audio-service/requirements.txt
essentia-tensorflow
flask
requests
yt-dlp
